feat-JSMP-12: created a separate component PopUpInfoPlace
