+++ src/app/components/Home/Office/Canvas/PopUpInfoPlace/components/index.ts
@@ -1,0 +1,19 @@
+import Button from "@material-ui/core/Button";
+import Dialog from "@material-ui/core/Dialog";
+import MuiDialogTitle from "@material-ui/core/DialogTitle";
+import MuiDialogContent from "@material-ui/core/DialogContent";
+import MuiDialogActions from "@material-ui/core/DialogActions";
+import IconButton from "@material-ui/core/IconButton";
+import CloseIcon from "@material-ui/icons/Close";
+import Typography from "@material-ui/core/Typography";
+
+export {
+  Button,
+  Dialog,
+  MuiDialogTitle,
+  MuiDialogContent,
+  MuiDialogActions,
+  IconButton,
+  CloseIcon,
+  Typography,
+};
