+++ src/app/components/Home/Office/Canvas/index.tsx
@@ -1,30 +1,18 @@
 import React, { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
 
-// selectors
-import { getDataFloors, getSelectedFloor } from "../../../../redux/selectors";
-// types
-import { Places } from "../../../../services/BD/type/Floors";
+// component
+import PopUpInfoPlace from "./PopUpInfoPlace";
 
 // style
 import { makeStyles } from "@material-ui/core/styles";
 import WrapperCanvas from "./components/WrapperCanvas";
 
-// material-ui
-import {
-  createStyles,
-  Theme,
-  withStyles,
-  WithStyles,
-} from "@material-ui/core/styles";
-import Button from "@material-ui/core/Button";
-import Dialog from "@material-ui/core/Dialog";
-import MuiDialogTitle from "@material-ui/core/DialogTitle";
-import MuiDialogContent from "@material-ui/core/DialogContent";
-import MuiDialogActions from "@material-ui/core/DialogActions";
-import IconButton from "@material-ui/core/IconButton";
-import CloseIcon from "@material-ui/icons/Close";
-import Typography from "@material-ui/core/Typography";
+// selectors
+import { getDataFloors, getSelectedFloor } from "../../../../redux/selectors";
+
+// types
+import { Places } from "../../../../services/BD/type/Floors";
 
 const useStyles = makeStyles({
   canvas: {
@@ -49,62 +37,6 @@ const useStyles = makeStyles({
     cursor: "pointer",
   },
 });
-
-const styles = (theme: Theme) =>
-  createStyles({
-    root: {
-      margin: 0,
-      padding: theme.spacing(2),
-    },
-    closeButton: {
-      position: "absolute",
-      right: theme.spacing(1),
-      top: theme.spacing(1),
-      color: theme.palette.grey[500],
-    },
-  });
-
-export interface DialogTitleProps extends WithStyles<typeof styles> {
-  id: string;
-  children: React.ReactNode;
-  onClose: () => void;
-}
-
-const DialogTitle = withStyles(styles)((props: DialogTitleProps) => {
-  const { children, classes, onClose, ...other } = props;
-  return (
-    <MuiDialogTitle
-      disableTypography
-      className={classes.root}
-      {...other}
-      style={{ width: "400px" }}
-    >
-      <Typography variant="h6">{children}</Typography>
-      {onClose ? (
-        <IconButton
-          aria-label="close"
-          className={classes.closeButton}
-          onClick={onClose}
-        >
-          <CloseIcon />
-        </IconButton>
-      ) : null}
-    </MuiDialogTitle>
-  );
-});
-
-const DialogContent = withStyles((theme: Theme) => ({
-  root: {
-    padding: theme.spacing(2),
-  },
-}))(MuiDialogContent);
-
-const DialogActions = withStyles((theme: Theme) => ({
-  root: {
-    margin: 0,
-    padding: theme.spacing(1),
-  },
-}))(MuiDialogActions);
 
 interface InfoAboutWorkplace {
   label: string;
@@ -175,10 +107,15 @@ const Canvas: React.FC = () => {
                 <text
                   id={`${index}`}
                   className={classes.placeLabel}
-                  x={item.placeStatus.coordinates.x + item.placeStatus.coordinates.width / 2 - 16}
+                  x={
+                    item.placeStatus.coordinates.x +
+                    item.placeStatus.coordinates.width / 2 -
+                    16
+                  }
                   y={
                     item.placeStatus.coordinates.y +
-                    item.placeStatus.coordinates.height / 2 + 4
+                    item.placeStatus.coordinates.height / 2 +
+                    4
                   }
                 >
                   {item.label}
@@ -188,38 +125,11 @@ const Canvas: React.FC = () => {
           </svg>
         </WrapperCanvas>
       ) : null}
-      <Dialog
-        onClose={handleClose}
-        aria-labelledby="customized-dialog-title"
-        open={open}
-      >
-        <DialogTitle id="customized-dialog-title" onClose={handleClose}>
-          {infoAboutWorkplace.label}
-        </DialogTitle>
-        <DialogContent dividers>
-          <Typography gutterBottom>Type: {infoAboutWorkplace.type}</Typography>
-          <Typography gutterBottom>
-            Occupant: {infoAboutWorkplace.occupant}
-          </Typography>
-          <Typography gutterBottom>
-            start: {infoAboutWorkplace.startReservation}
-          </Typography>
-          <Typography gutterBottom>
-            end: {infoAboutWorkplace.endReservation}
-          </Typography>
-        </DialogContent>
-        <DialogActions>
-          <Button
-            autoFocus
-            onClick={handleClose}
-            // color="secondary"
-            variant="contained"
-            color="primary"
-          >
-            Save changes
-          </Button>
-        </DialogActions>
-      </Dialog>
+      <PopUpInfoPlace
+        handleEvent={handleClose}
+        dataAboutWorkplace={infoAboutWorkplace}
+        visibility={open}
+      />
     </>
   );
 };
